+++ src/senaite/core/z3cform/widgets/address.py
@@ -235,9 +235,8 @@ class AddressWidget(HTMLFormElement, Widget):
                 subdivisions = geo.get_subdivisions(country, [])
                 sub1[country] = map(lambda sub: sub.name, subdivisions)
 
-                label = _("State")
-                if subdivisions:
-                    label = _(subdivisions[0].type)
+                state_label = _(u"label_subdivision_state", default=u"State")
+                label = geo.get_subdivision_type(subdivisions, state_label)
                 labels[country]["subdivision1"] = translate(label)
 
             subdivision1 = item.get("subdivision1")
@@ -245,9 +244,10 @@ class AddressWidget(HTMLFormElement, Widget):
                 subdivisions = geo.get_subdivisions(subdivision1, [])
                 sub2[subdivision1] = map(lambda sub: sub.name, subdivisions)
 
-                label = _("District")
-                if subdivisions:
-                    label = _(subdivisions[0].type)
+                district_label = _(
+                    u"label_subdivision_district", default=u"District"
+                )
+                label = geo.get_subdivision_type(subdivisions, district_label)
                 labels[country]["subdivision2"] = translate(label)
 
         attributes = {
@@ -305,12 +305,13 @@ class AjaxSubdivisions(BrowserView):
         # Extract the subdivisions for this parent
         parent = safe_unicode(parent)
         items = geo.get_subdivisions(parent, default=[])
+        subdivision_type = geo.get_subdivision_type(items, "State")
 
         def to_dict(subdivision):
             return {
                 "name": subdivision.name,
                 "code": subdivision.code,
-                "type": self.context.translate(_(subdivision.type)),
+                "type": self.context.translate(_(subdivision_type)),
             }
         return [to_dict(item) for item in items]